+++ libbpf-rs/CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
-- Added `KprobeMultiLinkInfo` fields that required a second info call.
+- Added `KprobeMultiLinkInfo` & `UprobeMultiLinkInfo` fields that required a
+  second info call.
 
 
 0.26.2